Move Timeline components to module
